Extract sync diagnose actions
