+++ dev-docs/module-reference.md
@@ -1390,7 +1390,8 @@ Not separately documented because their exports are best read from source — ke
 - `sync-environment.js` — relay URL selection, relay connectivity probing, and browser capability checks for Evolu sync.
 - `sync-identity.js` — BIP-39/QR lazy loaders plus mnemonic read/restore helpers for Evolu owner identity.
 - `sync-diagnostics.js` — Evolu row diagnostics snapshot and copy-text formatting for the Sync Diagnose modal.
-- `sync-diagnose-ui.js` — Sync Diagnose modal rendering, relay-storage actions, identity rotation UI, telemetry reset, and lean-sync cutover/backfill handlers.
+- `sync-diagnose-ui.js` — Sync Diagnose modal rendering, copy snapshot handling, and compatibility re-exports for Diagnose action handlers.
+- `sync-diagnose-actions.js` — Sync Diagnose relay-storage actions, identity rotation UI, telemetry reset, and lean-sync cutover/backfill handlers.
 - `sync-actions.js` — user-triggered sync actions, local storage cleanup, all-profile initial push, save debouncing, chat debouncing, profile-metadata push scheduling, and AI-setting push scheduling.
 - `sync-ui.js` — header sync badge, popover rendering, status subscription, and activity-log copy helpers.
 - `sync-payload.js` — Evolu wire-payload helpers; outbound profile/config/chat/display envelope assembly, Phase 2 cutover payload shape, gzip envelope handling, inbound payload parsing, and local-only data stripping for wearable credentials and SNP map rows.
